fix(hintLog): fix incorrect choice of error in line numbers
Previously, the findLineNumber function was not
outputing correct line numbers because
each module will have a different stack trace. This fix
adds a parameter so the module can call the correct line.
Additionally, the fix altered the logic of detecting the anonymous
tag that should be split out if it is present.